pyproject.toml
[project]
name = "dataviz-agent"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "streamlit",
    "pandas",
    "plotly",
    "openai",
    "python-dotenv"
]